Struct rusoto_xray::XRayClient

pub struct XRayClient { /* fields omitted */ }

A client for the AWS X-Ray API.

Methods

impl XRayClient

pub fn new(region: Region) -> XRayClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> XRayClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,