Crate rusoto_xray

AWS X-Ray provides APIs for managing debug traces and retrieving service maps and other data created by processing those traces.

If you're using the service, you're probably looking for XRayClient and XRay.

Structs

Alias
An alias for an edge.

AnnotationValue
Value of a segment annotation. Has one of three value types: Number, Boolean or String.

AvailabilityZoneDetail
A list of availability zones corresponding to the segments in a trace.

BackendConnectionErrors

BatchGetTracesRequest

BatchGetTracesResult

CreateGroupRequest

CreateGroupResult

CreateSamplingRuleRequest

CreateSamplingRuleResult

DeleteGroupRequest

DeleteGroupResult

DeleteSamplingRuleRequest

DeleteSamplingRuleResult

Edge
Information about a connection between two services.

EdgeStatistics
Response statistics for an edge.

EncryptionConfig
A configuration document that specifies encryption configuration settings.

ErrorRootCause
The root cause of a trace summary error.

ErrorRootCauseEntity
A collection of segments and corresponding subsegments associated to a trace summary error.

ErrorRootCauseService
A collection of fields identifying the services in a trace summary error.

ErrorStatistics
Information about requests that failed with a 4xx Client Error status code.

FaultRootCause
The root cause information for a trace summary fault.

FaultRootCauseEntity
A collection of segments and corresponding subsegments associated to a trace summary fault error.

FaultRootCauseService
A collection of fields identifying the services in a trace summary fault.

FaultStatistics
Information about requests that failed with a 5xx Server Error status code.

GetEncryptionConfigRequest

GetEncryptionConfigResult

GetGroupRequest

GetGroupResult

GetGroupsRequest

GetGroupsResult

GetSamplingRulesRequest

GetSamplingRulesResult

GetSamplingStatisticSummariesRequest

GetSamplingStatisticSummariesResult

GetSamplingTargetsRequest

GetSamplingTargetsResult

GetServiceGraphRequest

GetServiceGraphResult

GetTimeSeriesServiceStatisticsRequest

GetTimeSeriesServiceStatisticsResult

GetTraceGraphRequest

GetTraceGraphResult

GetTraceSummariesRequest

GetTraceSummariesResult

Group
Details and metadata for a group.

GroupSummary
Details for a group without metadata.

HistogramEntry
An entry in a histogram for a statistic. A histogram maps the range of observed values on the X axis, and the prevalence of each value on the Y axis.

Http
Information about an HTTP request.

InstanceIdDetail
A list of EC2 instance IDs corresponding to the segments in a trace.

PutEncryptionConfigRequest

PutEncryptionConfigResult

PutTelemetryRecordsRequest

PutTelemetryRecordsResult

PutTraceSegmentsRequest

PutTraceSegmentsResult

ResourceARNDetail
A list of resources ARNs corresponding to the segments in a trace.

ResponseTimeRootCause
The root cause information for a response time warning.

ResponseTimeRootCauseEntity
A collection of segments and corresponding subsegments associated to a response time warning.

ResponseTimeRootCauseService
A collection of fields identifying the service in a response time warning.

RootCauseException
The exception associated with a root cause.

SamplingRule
A sampling rule that services use to decide whether to instrument a request. Rule fields can match properties of the service, or properties of a request. The service can ignore rules that don't match its properties.

SamplingRuleRecord
A SamplingRule and its metadata.

SamplingRuleUpdate
A document specifying changes to a sampling rule's configuration.

SamplingStatisticSummary
Aggregated request sampling data for a sampling rule across all services for a 10 second window.

SamplingStatisticsDocument
Request sampling results for a single rule from a service. Results are for the last 10 seconds unless the service has been assigned a longer reporting interval after a previous call to GetSamplingTargets.

SamplingStrategy
The name and value of a sampling rule to apply to a trace summary.

SamplingTargetDocument
Temporary changes to a sampling rule configuration. To meet the global sampling target for a rule, X-Ray calculates a new reservoir for each service based on the recent sampling results of all services that called GetSamplingTargets.

Segment
A segment from a trace that has been ingested by the X-Ray service. The segment can be compiled from documents uploaded with PutTraceSegments, or an inferred segment for a downstream service, generated from a subsegment sent by the service that called it.

For the full segment document schema, see AWS X-Ray Segment Documents in the AWS X-Ray Developer Guide.

Service
Information about an application that processed requests, users that made requests, or downstream services, resources and applications that an application used.

ServiceId

ServiceStatistics
Response statistics for a service.

TelemetryRecord

TimeSeriesServiceStatistics
A list of TimeSeriesStatistic structures.

Trace
A collection of segment documents with matching trace IDs.

TraceSummary
Metadata generated from the segment documents in a trace.

TraceUser
Information about a user recorded in segment documents.

UnprocessedStatistics
Sampling statistics from a call to GetSamplingTargets that X-Ray could not process.

UnprocessedTraceSegment
Information about a segment that failed processing.

UpdateGroupRequest

UpdateGroupResult

UpdateSamplingRuleRequest

UpdateSamplingRuleResult

ValueWithServiceIds
Information about a segment annotation.

XRayClient
A client for the AWS X-Ray API.

Enums

BatchGetTracesError
Errors returned by BatchGetTraces

CreateGroupError
Errors returned by CreateGroup

CreateSamplingRuleError
Errors returned by CreateSamplingRule

DeleteGroupError
Errors returned by DeleteGroup

DeleteSamplingRuleError
Errors returned by DeleteSamplingRule

GetEncryptionConfigError
Errors returned by GetEncryptionConfig

GetGroupError
Errors returned by GetGroup

GetGroupsError
Errors returned by GetGroups

GetSamplingRulesError
Errors returned by GetSamplingRules

GetSamplingStatisticSummariesError
Errors returned by GetSamplingStatisticSummaries

GetSamplingTargetsError
Errors returned by GetSamplingTargets

GetServiceGraphError
Errors returned by GetServiceGraph

GetTimeSeriesServiceStatisticsError
Errors returned by GetTimeSeriesServiceStatistics

GetTraceGraphError
Errors returned by GetTraceGraph

GetTraceSummariesError
Errors returned by GetTraceSummaries

PutEncryptionConfigError
Errors returned by PutEncryptionConfig

PutTelemetryRecordsError
Errors returned by PutTelemetryRecords

PutTraceSegmentsError
Errors returned by PutTraceSegments

UpdateGroupError
Errors returned by UpdateGroup

UpdateSamplingRuleError
Errors returned by UpdateSamplingRule

Traits

XRay
Trait representing the capabilities of the AWS X-Ray API. AWS X-Ray clients implement this trait.