Crate rusoto_xray

AWS X-Ray provides APIs for managing debug traces and retrieving service maps and other data created by processing those traces.

If you're using the service, you're probably looking for XRayClient and XRay.

Structs

Alias
An alias for an edge.

AnnotationValue
Value of a segment annotation. Has one of three value types: Number, Boolean or String.

BackendConnectionErrors

BatchGetTracesRequest

BatchGetTracesResult

Edge
Information about a connection between two services.

EdgeStatistics
Response statistics for an edge.

EncryptionConfig
A configuration document that specifies encryption configuration settings.

ErrorStatistics
Information about requests that failed with a 4xx Client Error status code.

FaultStatistics
Information about requests that failed with a 5xx Server Error status code.

GetEncryptionConfigRequest

GetEncryptionConfigResult

GetServiceGraphRequest

GetServiceGraphResult

GetTraceGraphRequest

GetTraceGraphResult

GetTraceSummariesRequest

GetTraceSummariesResult

HistogramEntry
An entry in a histogram for a statistic. A histogram maps the range of observed values on the X axis, and the prevalence of each value on the Y axis.

Http
Information about an HTTP request.

PutEncryptionConfigRequest

PutEncryptionConfigResult

PutTelemetryRecordsRequest

PutTelemetryRecordsResult

PutTraceSegmentsRequest

PutTraceSegmentsResult

Segment
A segment from a trace that has been ingested by the X-Ray service. The segment can be compiled from documents uploaded with PutTraceSegments, or an inferred segment for a downstream service, generated from a subsegment sent by the service that called it.

For the full segment document schema, see AWS X-Ray Segment Documents in the AWS X-Ray Developer Guide.

Service
Information about an application that processed requests, users that made requests, or downstream services, resources and applications that an application used.

ServiceId

ServiceStatistics
Response statistics for a service.

TelemetryRecord

Trace
A collection of segment documents with matching trace IDs.

TraceSummary
Metadata generated from the segment documents in a trace.

TraceUser
Information about a user recorded in segment documents.

UnprocessedTraceSegment
Information about a segment that failed processing.

ValueWithServiceIds
Information about a segment annotation.

XRayClient
A client for the AWS X-Ray API.

Enums

BatchGetTracesError
Errors returned by BatchGetTraces

GetEncryptionConfigError
Errors returned by GetEncryptionConfig

GetServiceGraphError
Errors returned by GetServiceGraph

GetTraceGraphError
Errors returned by GetTraceGraph

GetTraceSummariesError
Errors returned by GetTraceSummaries

PutEncryptionConfigError
Errors returned by PutEncryptionConfig

PutTelemetryRecordsError
Errors returned by PutTelemetryRecords

PutTraceSegmentsError
Errors returned by PutTraceSegments

Traits

XRay
Trait representing the capabilities of the AWS X-Ray API. AWS X-Ray clients implement this trait.